+++ src/index.ts
@@ -2,6 +2,7 @@ export * from './decorators/ConnectedSocket';
 export * from './decorators/EmitOnFail';
 export * from './decorators/EmitOnSuccess';
 export * from './decorators/MessageBody';
+export * from './decorators/MessageAck';
 export * from './decorators/Middleware';
 export * from './decorators/NspParam';
 export * from './decorators/NspParams';
@@ -17,8 +18,10 @@ export * from './decorators/SocketRequest';
 export * from './decorators/SocketRooms';
 
 export * from './types/MiddlewareInterface';
+export * from './types/InterceptorInterface';
 export * from './types/TransformOptions';
 export * from './types/SocketControllersOptions';
 export * from './types/enums/SocketEventType';
+export * from './types/SocketEventContext';
 
 export * from './SocketControllers';