+++ src/types/SocketControllersOptions.ts
@@ -1,13 +1,15 @@
 import { Server } from 'socket.io';
 import { TransformOptions } from './TransformOptions';
-import { ScopedContainerGetterParams } from './ScopedContainerGetterParams';
+import { SocketEventContext } from './SocketEventContext';
 
 export interface SocketControllersOptions {
   container: { get<T>(someClass: { new (...args: any[]): T } | Function): T };
 
-  scopedContainerGetter?: (params: ScopedContainerGetterParams) => {
+  scopedContainerGetter?: (context: SocketEventContext) => {
     get<T>(someClass: { new (...args: any[]): T } | Function): T;
   };
+
+  scopedContainerDisposer?: (container: { get<T>(someClass: { new (...args: any[]): T } | Function): T }) => void;
 
   io?: Server;