+++ test/functional/scoped-controllers.spec.ts
@@ -2,13 +2,12 @@ import { createServer, Server as HttpServer } from 'http';
 import { Server } from 'socket.io';
 import { io, Socket } from 'socket.io-client';
 import { SocketControllers } from '../../src/SocketControllers';
-import { Container, Inject, Service, Token } from 'typedi';
+import { Container, ContainerInstance, Inject, Service, Token } from 'typedi';
 import { SocketController } from '../../src/decorators/SocketController';
 import { OnConnect } from '../../src/decorators/OnConnect';
 import { ConnectedSocket } from '../../src/decorators/ConnectedSocket';
 import { waitForEvent } from '../utilities/waitForEvent';
-import { EmitOnSuccess, OnMessage } from '../../src';
-import { ScopedContainerGetterParams } from '../../src/types/ScopedContainerGetterParams';
+import { EmitOnSuccess, OnMessage, SocketEventContext } from '../../src';
 
 describe('Scoped controllers', () => {
   const PORT = 8080;
@@ -73,7 +72,7 @@ describe('Scoped controllers', () => {
       io: wsApp,
       container: Container,
       controllers: [TestController],
-      scopedContainerGetter: (args: ScopedContainerGetterParams) => {
+      scopedContainerGetter: (args: SocketEventContext) => {
         return Container.of(Math.random().toString());
       },
     });
@@ -112,7 +111,7 @@ describe('Scoped controllers', () => {
       io: wsApp,
       container: Container,
       controllers: [TestController],
-      scopedContainerGetter: (args: ScopedContainerGetterParams) => {
+      scopedContainerGetter: (args: SocketEventContext) => {
         return Container.of(Math.random().toString());
       },
     });
@@ -154,7 +153,7 @@ describe('Scoped controllers', () => {
       io: wsApp,
       container: Container,
       controllers: [TestController],
-      scopedContainerGetter: (args: ScopedContainerGetterParams) => {
+      scopedContainerGetter: (args: SocketEventContext) => {
         const container = Container.of(counter.toString());
         container.set(token, counter);
         counter++;
@@ -189,7 +188,7 @@ describe('Scoped controllers', () => {
       io: wsApp,
       container: Container,
       controllers: [TestController],
-      scopedContainerGetter: (args: ScopedContainerGetterParams) => {
+      scopedContainerGetter: (args: SocketEventContext) => {
         testResult.push(args);
         return Container.of('');
       },
@@ -205,5 +204,53 @@ describe('Scoped controllers', () => {
     expect(testResult[1].eventName).toBe('test');
     expect(testResult[1].messageArgs).toEqual(['args1']);
     expect(testResult[1].nspParams).toEqual({ test: 'string' });
+  });
+
+  it('container should be disposed', async () => {
+    const token = new Token('ADDITIONAL');
+
+    @Service({ global: true })
+    class TestService {}
+
+    @SocketController('/string')
+    @Service()
+    class TestController {
+      constructor(private testService: TestService, @Inject(token) public myAdditional: number) {}
+
+      @OnConnect()
+      connected(@ConnectedSocket() socket: Socket) {
+        socket.emit('connected');
+      }
+
+      @OnMessage('test')
+      @EmitOnSuccess('done')
+      test() {
+        testResult.push(this.myAdditional);
+      }
+    }
+
+    let container;
+    socketControllers = new SocketControllers({
+      io: wsApp,
+      container: Container,
+      controllers: [TestController],
+      scopedContainerGetter: () => {
+        container = Container.of('test');
+        container.set(token, 'test');
+        return container;
+      },
+      scopedContainerDisposer: (scopedContainer: ContainerInstance) => {
+        scopedContainer.reset({ strategy: 'resetServices' });
+      },
+    });
+    wsClient = io(PATH_FOR_CLIENT + '/string', { reconnection: false, timeout: 5000, forceNew: true });
+
+    await waitForEvent(wsClient, 'connected');
+    wsClient.emit('test');
+    await waitForEvent(wsClient, 'done');
+
+    expect(Container.has(TestService)).toBe(true);
+    expect(container.has(token)).toBe(false);
+    expect(container.has(TestController)).toBe(false);
   });
 });